fixed some stuff

+++ src/main/java/nl/tudelft/jpacman/npc/Ghost.java
@@ -77,7 +77,7 @@ public abstract class Ghost extends Unit {
      * @return The suggested delay between moves in milliseconds.
      */
     public long getInterval() {
-        return this.moveInterval + new Random().nextInt(this.intervalVariation);
+        return (long)this.moveInterval + new Random().nextInt(this.intervalVariation);
     }
 
     /**
